# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "In Abstract Controller",
        "version": "1"
    },
    "paths": {
...
        "/users/cars": {
            "post": {
                "tags": [
                    "users"
                ],
                "operationId": "32134785f4984fdf23994e68ae7ca584",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCarOfferRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateCarOfferRequestData": {
                "properties": {
                    "manufacturer_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "is_new_car": {
                        "type": "integer"
                    },
                    "year_manufactured": {
                        "type": "integer",
                        "nullable": true
                    },
                    "car_color": {
                        "type": "integer",
                        "nullable": true
                    },
                    "model": {
                        "type": "integer",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "car_price": {
                        "type": "integer"
                    },
                    "car_sell_currency": {
                        "$ref": "#/components/schemas/Currency"
                    },
                    "fuel_type": {
                        "$ref": "#/components/schemas/FuelType"
                    },
                    "car_sell_location": {
                        "type": "integer"
                    },
                    "is_car_shippable_to_a_different_city": {
                        "type": "boolean"
                    },
                    "car_import_type": {
                        "$ref": "#/components/schemas/ImportType"
                    },
                    "car_label_origin": {
                        "type": "integer"
                    },
                    "transmission": {
                        "$ref": "#/components/schemas/TransmissionType"
                    },
                    "miles_travelled_in_km": {
                        "type": "integer",
                        "nullable": true
                    },
                    "has_tuf_check_passed": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "user_has_legal_car_papers": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "faragha_jahzeh": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "is_tajrobeh": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Currency": {
                "description": "[0 => Male, 1 => Female]",
                "type": "integer",
                "enum": [
                    0,
                    1
                ]
            },
            "FuelType": {
                "description": "[0 => Male, 1 => Female]",
                "type": "integer",
                "enum": [
                    0,
                    1,
                    2
                ]
            },
...
            "ImportType": {
                "description": "[0 => Male, 1 => Female]",
                "type": "integer",
                "enum": [
                    0,
                    1,
                    2,
                    3
                ]
            },
...
            "TransmissionType": {
                "description": "[0 => Male, 1 => Female]",
                "type": "integer",
                "enum": [
                    0,
                    1
                ]
            }
        },
...
        "securitySchemes": {
            "ref_to_be_used_below_csrf": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ref_to_be_used_below_csrf": []
        }
    ],
...
}